ver 0.008 -fixed bug for reshuffling big data. -fixed some comments (grammatic errors)
git-svn-id: svn+ssh://svn.r-forge.r-project.org/svnroot/genabel/pkg/OmicABEL@1262 edb9625f-4e0d-4859-8d74-9fd3b1da38cb

+++ src/reshuffle/reshuffle.h
@@ -28,7 +28,7 @@ public:
 	void write_data_chi(ifstream&,ofstream&);
 	void write_slim_data(ifstream&,ofstream&);
 	void write_herest(ifstream&,ofstream&);
-	int herest_startpos;
+	int64_t herest_startpos;
 	int est_shift(int);
 	int est_beta_shift(int);
 	void run(Parameters);
